chore: move channel management from CH reader to rule DB (#5934)
