+++ pkg/query-service/integrations/alertManager/manager.go
@@ -24,42 +24,62 @@ type Manager interface {
 	TestReceiver(receiver *Receiver) *model.ApiError
 }
 
-func New(url string) (Manager, error) {
+func defaultOptions() []ManagerOptions {
+	return []ManagerOptions{
+		WithURL(constants.GetAlertManagerApiPrefix()),
+		WithChannelApiPath(constants.AmChannelApiPath),
+	}
+}
 
-	if url == "" {
-		url = constants.GetAlertManagerApiPrefix()
+type ManagerOptions func(m *manager) error
+
+func New(opts ...ManagerOptions) (Manager, error) {
+	m := &manager{}
+
+	newOpts := defaultOptions()
+	newOpts = append(newOpts, opts...)
+
+	for _, opt := range newOpts {
+		err := opt(m)
+		if err != nil {
+			return nil, err
+		}
 	}
 
-	urlParsed, err := neturl.Parse(url)
-	if err != nil {
-		return nil, err
-	}
+	return m, nil
+}
 
-	return &manager{
-		url:       url,
-		parsedURL: urlParsed,
-	}, nil
+func WithURL(url string) ManagerOptions {
+	return func(m *manager) error {
+		m.url = url
+		parsedURL, err := neturl.Parse(url)
+		if err != nil {
+			return err
+		}
+		m.parsedURL = parsedURL
+		return nil
+	}
+}
+
+func WithChannelApiPath(path string) ManagerOptions {
+	return func(m *manager) error {
+		m.channelApiPath = path
+		return nil
+	}
 }
 
 type manager struct {
-	url       string
-	parsedURL *neturl.URL
+	url            string
+	parsedURL      *neturl.URL
+	channelApiPath string
 }
 
-func prepareAmChannelApiURL() string {
-	basePath := constants.GetAlertManagerApiPrefix()
-	AmChannelApiPath := constants.AmChannelApiPath
-
-	if len(AmChannelApiPath) > 0 && rune(AmChannelApiPath[0]) == rune('/') {
-		AmChannelApiPath = AmChannelApiPath[1:]
-	}
-
-	return fmt.Sprintf("%s%s", basePath, AmChannelApiPath)
+func (m *manager) prepareAmChannelApiURL() string {
+	return fmt.Sprintf("%s%s", m.url, m.channelApiPath)
 }
 
-func prepareTestApiURL() string {
-	basePath := constants.GetAlertManagerApiPrefix()
-	return fmt.Sprintf("%s%s", basePath, "v1/testReceiver")
+func (m *manager) prepareTestApiURL() string {
+	return fmt.Sprintf("%s%s", m.url, "v1/testReceiver")
 }
 
 func (m *manager) URL() *neturl.URL {
@@ -79,7 +99,7 @@ func (m *manager) AddRoute(receiver *Receiver) *model.ApiError {
 
 	receiverString, _ := json.Marshal(receiver)
 
-	amURL := prepareAmChannelApiURL()
+	amURL := m.prepareAmChannelApiURL()
 	response, err := http.Post(amURL, contentType, bytes.NewBuffer(receiverString))
 
 	if err != nil {
@@ -97,7 +117,7 @@ func (m *manager) AddRoute(receiver *Receiver) *model.ApiError {
 func (m *manager) EditRoute(receiver *Receiver) *model.ApiError {
 	receiverString, _ := json.Marshal(receiver)
 
-	amURL := prepareAmChannelApiURL()
+	amURL := m.prepareAmChannelApiURL()
 	req, err := http.NewRequest(http.MethodPut, amURL, bytes.NewBuffer(receiverString))
 
 	if err != nil {
@@ -126,7 +146,7 @@ func (m *manager) DeleteRoute(name string) *model.ApiError {
 	values := map[string]string{"name": name}
 	requestData, _ := json.Marshal(values)
 
-	amURL := prepareAmChannelApiURL()
+	amURL := m.prepareAmChannelApiURL()
 	req, err := http.NewRequest(http.MethodDelete, amURL, bytes.NewBuffer(requestData))
 
 	if err != nil {
@@ -156,7 +176,7 @@ func (m *manager) TestReceiver(receiver *Receiver) *model.ApiError {
 
 	receiverBytes, _ := json.Marshal(receiver)
 
-	amTestURL := prepareTestApiURL()
+	amTestURL := m.prepareTestApiURL()
 	response, err := http.Post(amTestURL, contentType, bytes.NewBuffer(receiverBytes))
 
 	if err != nil {
