+++ pkg/query-service/model/response.go
@@ -618,6 +618,7 @@ type AlertsInfo struct {
 	LogsBasedAlerts              int      `json:"logsBasedAlerts"`
 	MetricBasedAlerts            int      `json:"metricBasedAlerts"`
 	TracesBasedAlerts            int      `json:"tracesBasedAlerts"`
+	TotalChannels                int      `json:"totalChannels"`
 	SlackChannels                int      `json:"slackChannels"`
 	WebHookChannels              int      `json:"webHookChannels"`
 	PagerDutyChannels            int      `json:"pagerDutyChannels"`
